+++ CHANGELOG.rst
@@ -37,6 +37,8 @@ Change Log
 - [FIXED]: some bugs in the `GreedyAgent` and `TopologyGreedy`
 - [FIXED]: `Issue#220 <https://github.com/rte-france/Grid2Op/issues/220>`_ `flow_bus_matrix` did not took into
   account disconnected powerlines, leading to impossibility to compute this matrix in some cases.
+- [FIXED]: `Issue#223 <https://github.com/rte-france/Grid2Op/issues/223>`_ : now able to plot a grid even
+  if there is nothing controllable in grid2op present in it.
 - [ADDED] support for the "alarm operator" / "attention budget" feature
 - [ADDED] retrieval of the `max_step` (ie the maximum number of step that can be performed for the current episode)
   in the observation